+++ industry_reforged/__init__.py
@@ -1,4 +1,4 @@
 """Initialize the app"""
 
-__version__ = "V0.1.0b6"
+__version__ = "V0.1.0b7"
 __title__ = "Industry"